requirements.txt
# requirements.txt
click==8.1.3
colorama==0.4.6
Jinja2==3.1.2
PyYAML==6.0
python-slugify==8.0.1